GROOV-3795-Refactor-rake-task (pull request #1008)
GROOV-3795 | Refactor all rake task

* Refactor all rake tasks

* Minor changes

* fixed ss spec


Approved-by: Hasan Nadeem

+++ lib/tasks/delete_old_print_pdfs.rake
@@ -4,8 +4,12 @@ namespace :doo do
     desc 'Delete Pdfs earlier than 3 days'
   
     task delete_print_pdfs: :environment do
+      next if $redis.get('delete_pdfs_earlier_than_3_days')
+
+      $redis.set('delete_pdfs_earlier_than_3_days', true)
+      $redis.expire('delete_pdfs_earlier_than_3_days', 180)
       tenants = Tenant.all
-      tenants.each do |tenant|
+      tenants.find_each do |tenant|
         Apartment::Tenant.switch! tenant.name
         PrintPdfLink.where('created_at <= ?', 3.days.ago).each do |link|
             GroovS3.delete_object(link.url.gsub("#{ENV['S3_BASE_URL']}/", ''))
